Fix register value decoding: signed temps and missing enum types
- immergas_modbus.cpp: cast raw temp register through int16_t before
  applying scale so negative temperatures (outdoor, cold water) are
  decoded correctly instead of appearing as ~65000 °C values
- immergas_pdus.h: add IM_PDU_U32=6, IM_PDU_S32=7, IM_PDU_FLOAT32=8
  to the enum; the runtime switch in immergas_modbus.cpp already
  handles these types but they were missing from the generated header,
  causing compile errors

+++ components/immergas_modbus/immergas_modbus.cpp
@@ -243,7 +243,7 @@ void ImmergasModbus::poll_device(uint16_t slave, IM_Device *dev) {
 			float value = 0.0f;
 			switch (e.type) {
 				case IM_PDU_TEMP:
-					if (!sub.empty()) value = static_cast<float>(sub[0]) * e.scale;
+					if (!sub.empty()) value = static_cast<float>(static_cast<int16_t>(sub[0])) * e.scale;
 					break;
 				case IM_PDU_U16:
 					if (!sub.empty()) value = static_cast<float>(sub[0]);
